add business list entity
